re-merged resources with everything else
So at some point in the past, I thought it was a good idea to separate out the
resources from the rest of the files. Unfortunately, the "source" and "resource"
files are too coupled, and the difference in location is confusing.

If I do have resource management in the future, it will be based on packs that
can be added on to the main game, and splitting the main resources into a
resource pack would be a natural consequence.